+++ neurokit2/eeg/mne_data.py
@@ -1,12 +1,15 @@
 # -*- coding: utf-8 -*-
 
-def mne_data(what="raw"):
+def mne_data(what="raw", path=None):
     """Utility function to easily access MNE datasets
 
     Parameters
     -----------
     what : str
         Can be 'raw' or 'filt-0-40_raw' (a filtered version).
+    path : str
+        Defaults to None, assuming that the MNE data folder already exists. If not,
+        specify the directory to download the folder.
 
     Returns
     -------
@@ -32,7 +35,12 @@ def mne_data(what="raw"):
     old_verbosity_level = mne.set_log_level(verbose="WARNING", return_old_level=True)
 
     if what in ["raw", "filt-0-40_raw"]:
-        path = mne.datasets.sample.data_path()
+        if path is None:
+            try:
+                path = mne.datasets.sample.data_path()
+            except ValueError:
+                raise ValueError("NeuroKit error: the mne sample data folder does not exist. ",
+                                 "Please specify a path to download the mne datasets.")
         path += '/MEG/sample/sample_audvis_' + what + '.fif'
         data = mne.io.read_raw_fif(path, preload=True)
         data = data.pick_types(meg=False, eeg=True)